Create mclip.py
Chapter 06 chapter project mclip.py - A multi-clipboard program using pre-defined keyphrases and copying them to the clipboard for use outside of the program.  Uses the sample text from the chapter but imports the pyperclip module it calls in the program so that it functions correctly.  **The pyperclip python module must be installed in order for this program to function.