Update threadedDownloadXkcd.py
Improved handling of "comicUrl" to ensure it includes the protocol.  Changed the loop range to 'range(startComic, endComic + 1)' to include the last comic in the range.
Added error handling to catch and print any 'requests' exceptions during the download process.
used a context manager ('with' statement) for opening and writing the image file to ensure proper closure.